refactor: create staging feature directory (Phase 5b)
Move Staging component and useStagingDragInteraction hook to a new
feature-based directory structure:

- src/features/staging/components/Staging.tsx
- src/features/staging/hooks/useStagingDragInteraction.ts
- Barrel exports at feature and component levels

Re-export stubs maintain backward compatibility at old locations:
- src/components/Staging.tsx
- src/hooks/interactions/useStagingDragInteraction.ts

These re-export stubs will be removed in Phase 6.

Part of folder structure reorganization Phase 5.
