+++ src/features/staging/components/index.ts
@@ -1,0 +1,2 @@
+// Barrel export for staging components
+export { Staging } from './Staging';
