+++ src/features/staging/hooks/index.ts
@@ -1,0 +1,2 @@
+// Barrel export for staging hooks
+export { useStagingDragInteraction } from './useStagingDragInteraction';
