+++ src/features/staging/index.ts
@@ -1,0 +1,3 @@
+// Barrel export for staging feature
+export * from './components';
+export * from './hooks';
